requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0

sqlalchemy>=2.0.29
alembic>=1.13.1
requests>=2.31.0

python-jose>=3.3.0
passlib>=1.7.4

# ⚠️ bcrypt precisa ser versão NOVA
bcrypt>=4.1.3

python-multipart>=0.0.9

pydantic>=2.6.4
pydantic-settings>=2.2.1

python-dotenv>=1.0.1
pypdf
apscheduler 
python-docx
openai
numpy 
psycopg2-binary
 